tools/flake8-torch/pyproject.toml
[project]
name = "flake8-torch"
dynamic = ["version"]
dependencies = ["flake8>=3.8.2", "PyYAML", "libcst==0.4.9"]

[project.entry-points]
"flake8.extension" = {TOR = "flake8_torch.checker:TorchChecker"}

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
pythonpath = [
  "."
]

[tool.black]
exclude = "tests/fixtures/*"

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.setuptools.dynamic]
version = {attr = "flake8_torch.checker.__version__"}
